fix for x86 build

+++ linker.cpp
@@ -2,7 +2,7 @@
 
 using namespace rrl;
 
-uintptr_t Linker::resolve_internal_symbol(std::string const &library, std::string const &symbol) const {
+uint64_t Linker::resolve_internal_symbol(std::string const &library, std::string const &symbol) const {
     if (auto it = libraries_.find(library); it != libraries_.end())
         return it->second.get_symbol_address(symbol);
     return 0;
